+++ src/coord/rmr/conn.h
@@ -64,10 +64,10 @@ typedef struct MRConn MRConn;
 typedef struct {
   dict *map;
   int nodeConns;
-  long long commandTimeoutMS;  // Timeout for commands sent on connections (0 = no timeout)
+  uint32_t commandTimeoutMS;  // Timeout for commands sent on connections (0 = no timeout)
 } MRConnManager;
 
-void MRConnManager_Init(MRConnManager *mgr, int nodeConns, long long commandTimeoutMS);
+void MRConnManager_Init(MRConnManager *mgr, int nodeConns, uint32_t commandTimeoutMS);
 
 /*
  * Gets the stateDict filled with connection pool states of different IORuntimes and
